demo/go/leetcode/14. 最长公共前缀_test.go
package leetcode

import (
	"strings"
	"testing"
)

func Test_longestCommonPrefix(t *testing.T) {
	type args struct {
		strs []string
	}
	tests := []struct {
		args args
		want string
	}{
		{
			args: args{
				strs: []string{"flower", "flow", "flight"},
			},
			want: "fl",
		},
		{
			args: args{
				strs: []string{"aflower", "bflow", "cflight"},
			},
			want: "",
		},
		{
			args: args{
				strs: []string{"dog", "racecar", "car"},
			},
			want: "",
		},
	}
	for _, tt := range tests {
		t.Run(strings.Join(tt.args.strs, "-"), func(t *testing.T) {
			if got := longestCommonPrefix(tt.args.strs); got != tt.want {
				t.Errorf("longestCommonPrefix() = %v, want %v", got, tt.want)
			}
		})
	}
}
